Move Wrapper to own File
Move RegexDateTimeWrapper class and interface to their own file for other codes using them
